requirements.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.26
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langsmith>=0.4.10
langchain-experimental>=0.0.50
langchain-mcp-adapters==0.1.14

# LLM Providers
langchain-openai>=0.0.8
langchain-anthropic>=0.1.0
langchain-aws>=0.1.0

# Vector Databases
chromadb>=0.4.22
pinecone-client>=3.0.0

# Graph Databases
neo4j==6.1.0

# Core Dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0
aiohttp>=3.9.0
mcp>=1.22.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
numexpr>=2.11.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
pysqlite3-binary>=0.5.4

# Jupyter for Interactive Development
jupyter>=1.0.0
ipykernel>=6.25.0